+++ src/app/models/user.model.ts
@@ -4,6 +4,7 @@ export class User {
   id?: number;
   name?: string;
   email: string;
+  emailVerified?: boolean;
   password?: string;
   token?: string;
   role?: string;